chore: Extract headers retrieval logic into separate module
